+++ src/components/stable/gux-listbox/gux-listbox.tsx
@@ -199,8 +199,10 @@ export class GuxListbox {
       this.listboxOptions.filter(listboxOption => !listboxOption.filtered)
         .length === 0;
 
-    if (this.filter) {
+    if (!this.allListboxOptionsFiltered && this.filter) {
       setFirstOptionActive(this.root);
+    } else {
+      clearActiveOptions(this.root);
     }
   }